+++ lib/python/CHANGELOG.md
@@ -2,6 +2,15 @@ VMware Aria Operations Integration SDK Library
 ----------------------------------------------
 
 ## Unreleased
+* If an AdapterDefinition is passed into a 'CollectResult', external object types
+  are no longer returned in the result unless they contain content. Relationships
+  are not affected. (Note: requires unreleased mp-test update)
+* Adds 'RelationshipUpdateModes' enum to control how relationships are returned from 
+  'CollectResult'
+* Add 'get_object' and 'get_objects_by_type' helper methods in 'CollectResult'
+* Add 'adapter_type()' and 'object_type()' helper methods in 'Object'
+* Add 'has_content()' method in 'Object' that returns true if the object contains 
+  at least one metric, property, or event
 * Add 'collection_number' to Adapter Instance. Starts at 0 and increments each time 
   'collect' is called. Requires server based off of base-python-adapter 0.10.0 or 
   higher.